refactor the code to gen3-qa architecture
