+++ codecept.conf.js
@@ -79,7 +79,7 @@ exports.config = {
     guppy: './services/apis/guppy/guppyService.js',
     mds: './services/apis/mds/mdsService.js',
     auditService: './services/apis/auditService/auditService.js',
-
+    nondbgap: './services/apis/nondbgapUsersync/nonDbgapService.js',
     // Pages
     home: './services/portal/home/homeService.js',
     indexing: './services/portal/indexing/indexingService.js',
